requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SocketIO==5.4.1
h11==0.14.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pkg_resources==0.0.0
pyserial==3.5
python-engineio==4.10.1
python-socketio==5.11.4
simple-websocket==1.1.0
Werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
